split the shim package
